+++ prompt_styles/dark-col-double-line.sh
@@ -82,13 +82,13 @@ fancygit_prompt_builder() {
         has_unpushed_commits=""
     fi
 
-    if fg_show_time
+    if fancygit_config_is "show-time" "true"
     then
         formatted_time=$(date +"${time_format}")
         prompt_time="${time}[${formatted_time}] ${time_end}"
     fi
 
-    if fg_show_user_at_machine
+    if fancygit_config_is "show-user-at-machine" "true"
     then
         user_at_host="${user}\\u${at}@${host}\\h "
         user_at_host_end="${bold_none}${bg_none}${s_darkgray01_bgdarkgray}"
@@ -112,7 +112,7 @@ fancygit_prompt_builder() {
     fi
 
     path_sign="\\W"
-    if fg_show_full_path
+    if fancygit_config_is "show-full-path" "true"
     then
         path_sign="\\w"
     fi