+++ prompt_styles/human-dark-single-line.sh
@@ -35,7 +35,7 @@ fancygit_prompt_builder() {
     local user_at_host=""
     local prompt_time=""
 
-    if fg_show_full_path
+    if fancygit_config_is "show-full-path" "true"
     then
         where="\w"
     fi
@@ -52,13 +52,13 @@ fancygit_prompt_builder() {
         branch_name=" ${red}on${none} ${orange}$branch_name${none}"
     fi
 
-    if fg_show_time
+    if fancygit_config_is "show-time" "true"
     then
       formatted_time=$(date +"${time_format}")
       prompt_time="[${formatted_time}] "
     fi
 
-    if fg_show_user_at_machine
+    if fancygit_config_is "show-user-at-machine" "true"
     then
         user_at_host="${user} ${at} ${host} ${in} "
     fi