+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change `fancygit update` command to `fancygit --update`.
 - Change `fancygit --colors-set` command to `fancygit --colors-apply`.
 - Change `fancygit configure-fonts` command to `fancygit --fonts-install`.
+- Change all fancy double-line themes. Now they show de user symbol at the same line.
 
 ### v6.4.8